Change pascal case to camel case
