+++ src/components/select/CustomSelect.jsx
@@ -38,6 +38,7 @@ const CustomSelect = ({ options, value, onChange }) => {
     return () => {
       document.removeEventListener("mousedown", handleClickOutside);
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
