+++ src/main.jsx
@@ -8,7 +8,7 @@ import { TemplateProvider } from "./middleware/resume";
 
 import en from "javascript-time-ago/locale/en.json";
 import ru from "javascript-time-ago/locale/ru.json";
-import { WalkthroughProvider } from "./context/WalkthroughContext";
+import { WalkthroughProvider } from "./middleware/walkthrough";
 
 TimeAgo.addDefaultLocale(en);
 TimeAgo.addLocale(ru);
