+++ src/pages/DefaultRoutes/resume/resume-landing/tips.jsx
@@ -5,7 +5,7 @@ import {
   ResumeTemplateSampleImage,
 } from "../../../../assets";
 import { ResumeButton } from "../../../../components/button";
-import InfiniteLooper from "../../../../components/elementScroll";
+import InfiniteLooper from "../../../../components/infiniteScroller";
 import "./index.css";
 import { slideIn, textVariantUp } from "../../../../utils/motion";
 import SectionWrapper from "../../../../components/animation/sectionWrapper";
